requirements.txt
langchain
gradio
faiss-cpu
pytesseract
pillow
unstructured
pdfminer.six
beautifulsoup4
requests
huggingface-hub
